added some stuff

+++ src/include/timer.h
@@ -11,4 +11,4 @@
 
 void init_timer(u32 freq);
 u32 get_ticks();
-void sleep(u32 sec);
+void sleep(double sec);
